+++ src/endpoints-factory.ts
@@ -1,8 +1,10 @@
 import {z} from 'zod';
+import {ApiResponse} from './api-response';
 import {Endpoint, Handler} from './endpoint';
-import {ApiResponse, FlatObject, IOSchema, Merge} from './helpers';
+import {FlatObject, IOSchema, Merge} from './helpers';
 import {Method, MethodsDefinition} from './method';
 import {MiddlewareDefinition} from './middleware';
+import {mimeJson, mimeMultipart} from './mime';
 import {defaultResultHandler, ResultHandlerDefinition} from './result-handler';
 
 type BuildProps<IN extends IOSchema, OUT extends IOSchema, mIN, mOUT, M extends Method> = {
@@ -10,6 +12,7 @@ type BuildProps<IN extends IOSchema, OUT extends IOSchema, mIN, mOUT, M extends 
   output: OUT;
   handler: Handler<z.output<Merge<IN, mIN>>, z.input<OUT>, mOUT>;
   description?: string;
+  type?: 'json' | 'upload'; // @todo can we detect the usage of z.upload() within input?
 } & MethodsDefinition<M>;
 
 /** mIN, mOUT - accumulated from all middlewares */
@@ -39,7 +42,7 @@ export class EndpointsFactory<mIN, mOUT, POS extends ApiResponse, NEG extends Ap
   }
 
   public build<IN extends IOSchema, OUT extends IOSchema, M extends Method>({
-    input, output, handler, description, ...rest
+    input, output, handler, description, type, ...rest
   }: BuildProps<IN, OUT, mIN, mOUT, M>) {
     return new Endpoint<IN, OUT, mIN, mOUT, M, POS, NEG>({
       handler, description,
@@ -47,6 +50,7 @@ export class EndpointsFactory<mIN, mOUT, POS extends ApiResponse, NEG extends Ap
       inputSchema: input,
       outputSchema: output,
       resultHandler: this.resultHandler,
+      mimeTypes: type === 'upload' ? [mimeMultipart] : [mimeJson],
       ...rest
     });
   }