+++ tests/system/example.spec.ts
@@ -1,7 +1,10 @@
 import fetch from 'node-fetch';
 import {spawn, ChildProcessWithoutNullStreams} from 'child_process';
+import {mimeMultipart} from '../../src/mime';
 import {waitFor} from '../helpers';
 import crypto from 'crypto';
+import FormData from 'form-data';
+import {readFileSync} from 'fs';
 
 describe('Example', () => {
   let example: ChildProcessWithoutNullStreams;
@@ -109,6 +112,27 @@ describe('Example', () => {
       expect(response.headers.get('Transfer-encoding')).toBe('chunked');
       const hash = crypto.createHash('sha1').update(await response.text()).digest('hex');
       expect(hash).toMatchSnapshot();
+    });
+
+    test('Should upload the file', async () => {
+      const filename = 'logo.svg';
+      const logo = readFileSync(filename, 'utf-8');
+      const data = new FormData();
+      data.append('avatar', logo, { filename });
+      data.append('str', 'test string value');
+      data.append('num', 123);
+      data.append('arr[0]', 456);
+      data.append('arr[1]', 789);
+      data.append('obj[some]', 'thing');
+      const response = await fetch('http://localhost:8090/v1/upload', {
+        method: 'POST',
+        headers: {
+          'Content-Type': `${mimeMultipart}; boundary=${data.getBoundary()}`,
+        },
+        body: data
+      });
+      const json = await response.json();
+      expect(json).toMatchSnapshot();
     });
   });