+++ tools/generate-license.ts
@@ -62,6 +62,11 @@ const libs: Lib[] = [
     name: 'Mime',
     url: 'https://github.com/broofa/mime',
     module: 'mime'
+  },
+  {
+    name: 'Express-FileUpload',
+    url: 'https://github.com/richardgirges/express-fileupload',
+    module: 'express-fileupload'
   }
 ];